Update botsvsbots.py
shaved a bit of lines off. 
stickybotshielded is now a normal pro bot, so only pro bot option need to be on for it to spawn
this is cuz bouncybotsemilite also spawns normally (= count as normal bot), and imo this makes it more consistent.
and slow mo will have epic music again. cu I realised sometimes a match could go on longer than I think.